+++ MathExpressionEvaluator/NodeFunctionCall.cs
@@ -14,9 +14,9 @@ namespace MathExpressionEvaluator
             Arguments = arguments;
         }
 
-        public override float? Eval(IContext context)
+        public override double? Eval(IContext context)
         {
-            var argumentValues = new float?[Arguments.Length];
+            var argumentValues = new double?[Arguments.Length];
             for (int i = 0; i < Arguments.Length; i++)
             {
                 argumentValues[i] = Arguments[i].Eval(context);
